Фикс замечаний. Обновил тесты
