+++ src/components/main/main.test.js
@@ -2,27 +2,8 @@ import React from "react";
 import renderer from "react-test-renderer";
 import Main from "./main.jsx";
 
+import {MOVIE_TITLES} from "../../consts.js";
 
-const MOVIE_TITLES = [
-  `Fantastic Beasts: The Crimes of Grindelwald`,
-  `Bohemian Rhapsody`,
-  `Macbeth`,
-  `Aviator`,
-  `We need to talk about Kevin`,
-  `What We Do in the Shadows`,
-  `Revenant`,
-  `Johnny English`,
-  `Shutter Island`,
-  `Pulp Fiction`,
-  `No Country for Old Men`,
-  `Snatch`, `Moonrise Kingdom`,
-  `Seven Years in Tibet`,
-  `Midnight Special`, `War of the Worlds`,
-  `Dardjeeling Limited`,
-  `Orlando`,
-  `Mindhunter`,
-  `Midnight Special`
-];
 
 describe(`Main Component`, () => {
   it(`Should main render correctly`, () => {
