+++ src/consts.js
@@ -1,0 +1,20 @@
+export const MOVIE_TITLES = [
+  `Fantastic Beasts: The Crimes of Grindelwald`,
+  `Bohemian Rhapsody`,
+  `Macbeth`,
+  `Aviator`,
+  `We need to talk about Kevin`,
+  `What We Do in the Shadows`,
+  `Revenant`,
+  `Johnny English`,
+  `Shutter Island`,
+  `Pulp Fiction`,
+  `No Country for Old Men`,
+  `Snatch`, `Moonrise Kingdom`,
+  `Seven Years in Tibet`,
+  `Midnight Special`, `War of the Worlds`,
+  `Dardjeeling Limited`,
+  `Orlando`,
+  `Mindhunter`,
+  `Midnight Special`
+];
