+++ src/index.js
@@ -2,39 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
 
-const promoInfo = {
-  title: `The Grand Budapest Hotel`,
-  genre: `Drama`,
-  year: 2014,
-};
-
-const MOVIE_TITLES = [
-  `Fantastic Beasts: The Crimes of Grindelwald`,
-  `Bohemian Rhapsody`,
-  `Macbeth`,
-  `Aviator`,
-  `We need to talk about Kevin`,
-  `What We Do in the Shadows`,
-  `Revenant`,
-  `Johnny English`,
-  `Shutter Island`,
-  `Pulp Fiction`,
-  `No Country for Old Men`,
-  `Snatch`, `Moonrise Kingdom`,
-  `Seven Years in Tibet`,
-  `Midnight Special`, `War of the Worlds`,
-  `Dardjeeling Limited`,
-  `Orlando`,
-  `Mindhunter`,
-  `Midnight Special`
-];
 
 ReactDOM.render(
-    <App
-      promoTitle={promoInfo.title}
-      promoGenre={promoInfo.genre}
-      promoYear={promoInfo.year}
-      movieTitles={MOVIE_TITLES}
-    />,
+    <App />,
     document.querySelector(`#root`)
 );
